refactor: 수익률 값 타입 수정

+++ src/main/java/lotto/cotroller/Controller.java
@@ -67,9 +67,9 @@ public class Controller {
         purchaseAmount = new PurchaseAmount(inputToInt);
     }
 
-    private int getEarningRate() {
-        int totalPrize = service.getTotalPrize(score);
-        int amount = purchaseAmount.getAmount();
+    private double getEarningRate() {
+        double totalPrize = service.getTotalPrize(score);
+        double amount = purchaseAmount.getAmount();
         return totalPrize / amount;
     }
 }
